various data type improvements
* fixes at least <FS>, which has been broken since b5325e00c402ec18034da4b4a4aaaefa87bb1fef.
* but probably many other commands as well in the DOS version when working with
  "very large" documents
* now compiles cleanly with -Wsign-compare

+++ tecdisp.c
@@ -2394,7 +2394,7 @@ struct buff_header *hbp = wptr->win_buffer;
 
     sbp = format_line_free_list;
     if(sbp != NULL){
-	if(sbp->fmt_buffer_size == term_columns){
+	if(sbp->fmt_buffer_size == (size_t)term_columns){
 	    format_line_free_list = sbp->fmt_next_line;
 	}
 	else {
